Added feature to combat system, monster damage is now modified by armor-percentage. Updated calculateStats to go calc armor bonuses as well. Changed various other things.

+++ src/interfaces/ItemInterface.java
@@ -6,9 +6,9 @@ public interface ItemInterface {
     
     public boolean isConsumable();
     
-    public int getDamageModifier();
+    public double getDamageModifier();
     
-    public int getHealthModifier();
+    public double getHealthModifier();
     
     public boolean isEquippable();
     
@@ -20,7 +20,7 @@ public interface ItemInterface {
     
     public void setKeyId(String roomName);
     
-    public int getArmorRating();
+    public double getArmorRating();
 
     
 }
